requirements.txt
asgiref==3.7.2
Django==4.2.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pytz==2023.4
sqlparse==0.4.4
django-cors-headers==3.13.0
